Renaming parser source code files to be more general
