+++ lib/parsing/error_handler.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "domain.hpp"
+#include "api.hpp"
 
 #include <boost/spirit/home/x3/support/ast/position_tagged.hpp>
 #include <boost/spirit/home/x3/support/utility/error_reporting.hpp>
