+++ test/test_parser.cpp
@@ -9,7 +9,7 @@
 
 #include "parsing/ast.hpp"
 #include "parsing/ast_adapted.hpp"
-#include "parsing/domain.hpp"
+#include "parsing/api.hpp"
 #include "parsing/parse.hpp"
 #include "util.h"
 #include <boost/optional.hpp>
