+++ test/test_unification.cpp
@@ -15,7 +15,7 @@
 
 #include "parsing/ast.hpp"
 #include "parsing/ast_adapted.hpp"
-#include "parsing/domain.hpp"
+#include "parsing/api.hpp"
 #include "parsing/parse.hpp"
 #include "util.h"
 #include "boost/optional.hpp"
